+++ test/processors/context.test.js
@@ -1,38 +1,10 @@
-const { strictEqual: eq, match } = require("assert");
+const { strictEqual: eq, ok, match } = require("assert");
 const {
   processors: { context },
 } = require("../..");
 
 describe("context", () => {
-  describe("message is an instance of error", () => {
-    it("should nest the error on the record", () => {
-      const fn = context();
-      const message = new Error("Oooh, Demons!");
-      const result = fn({ message });
-
-      eq(result.error.message, "Oooh, Demons!");
-      match(result.error.stack, /^Error: Oooh, Demons!/);
-    });
-
-    it("should set the record message", () => {
-      const fn = context();
-      const message = new Error("Oooh, Demons!");
-      const result = fn({ message });
-
-      eq(result.message, "Oooh, Demons!");
-    });
-
-    it("should nest with a custom property", () => {
-      const fn = context({ errorField: "e" });
-      const message = new Error("Oooh, Demons!");
-      const result = fn({ message });
-
-      eq(result.e.message, "Oooh, Demons!");
-      match(result.e.stack, /^Error: Oooh, Demons!/);
-    });
-  });
-
-  describe("context is an instance of error", () => {
+  describe("context is an Error", () => {
     it("should nest the error on the record", () => {
       const fn = context();
       const ctx = new Error("Oooh, Demons!");
@@ -52,7 +24,31 @@ describe("context", () => {
     });
   });
 
-  describe("context includes an instance of error", () => {
+  describe("context is an Array", () => {
+    it("should nest the array on the record", () => {
+      const fn = context();
+      const ctx = [1, 2, 3];
+      const result = fn({ ctx });
+
+      eq(result.items.length, 3);
+      ok(result.items[0], 1);
+      ok(result.items[1], 2);
+      ok(result.items[2], 3);
+    });
+
+    it("should nest with a custom property", () => {
+      const fn = context({ arrayField: "things" });
+      const ctx = [1, 2, 3];
+      const result = fn({ ctx });
+
+      eq(result.things.length, 3);
+      ok(result.things[0], 1);
+      ok(result.things[1], 2);
+      ok(result.things[2], 3);
+    });
+  });
+
+  describe("context is an Object", () => {
     it("should correctly capture error details when provided via context", () => {
       const fn = context();
       const ctx = { a: "b", error: new Error("Oooh, Demons!"), x: "y" };