+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to Semantic Versioning](https://semver.org/spec/v2.0.0.
 ## 3.0.0
 
 - Port the error processor functionality to the context processor and deprecate the error processor. See https://github.com/acuminous/tripitaka/issues/8
-- Add empty processor
-- Switch to prettier
+- Support Array contexts
+- Add empty processor for reporting missing messages
+- Introduce to prettier
 
 ## 2.1.7